fix hid demo build on linux

+++ src/xenia/hid/hid_demo.cc
@@ -15,7 +15,6 @@
 #include "xenia/base/clock.h"
 #include "xenia/base/logging.h"
 #include "xenia/base/main.h"
-#include "xenia/base/platform_win.h"
 #include "xenia/base/threading.h"
 #include "xenia/hid/input_system.h"
 #include "xenia/ui/gl/gl_provider.h"
@@ -25,6 +24,8 @@
 // Available input drivers:
 #include "xenia/hid/nop/nop_hid.h"
 #if XE_PLATFORM_WIN32
+#include "xenia/base/platform_win.h"
+
 #include "xenia/hid/winkey/winkey_hid.h"
 #include "xenia/hid/xinput/xinput_hid.h"
 #endif  // XE_PLATFORM_WIN32
@@ -177,8 +178,10 @@ void DrawUserInputStatus(uint32_t user_index) {
   ImGui::Text("     Shoulders: [%c][%c]",
               gamepad.buttons & X_INPUT_GAMEPAD_LEFT_SHOULDER ? 'L' : ' ',
               gamepad.buttons & X_INPUT_GAMEPAD_RIGHT_SHOULDER ? 'R' : ' ');
-  ImGui::Text("  Left Trigger: %3u", gamepad.left_trigger);
-  ImGui::Text(" Right Trigger: %3u", gamepad.right_trigger);
+  ImGui::Text("  Left Trigger: %3u",
+              static_cast<uint8_t>(gamepad.left_trigger));
+  ImGui::Text(" Right Trigger: %3u",
+              static_cast<uint8_t>(gamepad.right_trigger));
   ImGui::Text("    Left Thumb: %6d, %6d",
               static_cast<int32_t>(gamepad.thumb_lx),
               static_cast<int32_t>(gamepad.thumb_ly));
